Ajuste al crud de Softskill y pequeño ajuste en Experiencia laboral

+++ mi-portfolio/src/main/java/backendportfolio/miportfolio/controller/SoftskillController.java
@@ -1,10 +1,14 @@
 
 package backendportfolio.miportfolio.controller;
 
+import backendportfolio.miportfolio.dto.dtoSoftskill;
 import backendportfolio.miportfolio.entity.Softskill;
 import backendportfolio.miportfolio.interfaces.ISoftskillService;
 import java.util.List;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,40 +27,66 @@ public class SoftskillController {
     private ISoftskillService skillService;
     
     @GetMapping ("/skill/obtener")
-    public List<Softskill> getSoftskill(){
-        return skillService.getSoftskill();
+    public ResponseEntity<List<Softskill>> getSoftskill(){
+        List<Softskill> listSoftskill = skillService.getSoftskill();
+        return new ResponseEntity(listSoftskill, HttpStatus.OK);
     }
     
     @PostMapping ("/skill/crear")
-    public String agregarSoftskill(@RequestBody Softskill skill){
-        skillService.saveSoftskill(skill);
-        return "Se ha creado correctamente";
+    public ResponseEntity<?> agregarSoftskill(@RequestBody dtoSoftskill dtoSoftskill){
+        
+        if(StringUtils.isBlank(dtoSoftskill.getNombre()))
+            return new ResponseEntity(new Mensaje("El nombre es obligatorio"), HttpStatus.BAD_REQUEST);
+        
+        Softskill softskill = new Softskill(dtoSoftskill.getNombre(), dtoSoftskill.getNivel());
+        skillService.saveSoftskill(softskill);
+        
+        return new ResponseEntity(new Mensaje("Softskill agregada"), HttpStatus.OK);
     }
     
     @DeleteMapping ("/skill/borrar/{id}")
-    public String eliminarSoftskill(@PathVariable Long id){
+    public ResponseEntity<?> eliminarSoftskill(@PathVariable("id") Long id){
+        
+        if(!skillService.existsById(id))
+            return new ResponseEntity(new Mensaje("El ID no existe"), HttpStatus.BAD_REQUEST);   
+        
         skillService.deleteSoftskill(id);
-        return "Se ha eliminado correctamente";
+        
+        return new ResponseEntity(new Mensaje("Softskill eliminada"), HttpStatus.OK);
     }
     
     @PutMapping ("/skill/editar/{id}")
-    public Softskill editarSoftskill(@PathVariable Long id,
-                                    @RequestParam ("nombre") String nuevoNombre,
-                                    @RequestParam ("nivel") int nuevoNivel){
+    public ResponseEntity<?> editarSoftskill(@PathVariable("id") Long id, @RequestBody dtoSoftskill dtoSoftskill){
         
-        Softskill skill = skillService.findSoftskill(id);
+        if(!skillService.existsById(id))
+            return new ResponseEntity(new Mensaje("El ID no existe"), HttpStatus.BAD_REQUEST);
+        if(StringUtils.isBlank(dtoSoftskill.getNombre()))
+            return new ResponseEntity(new Mensaje("El nombre es obligatorio"), HttpStatus.BAD_REQUEST);
         
-        skill.setNombre(nuevoNombre);
-        skill.setNivel(nuevoNivel);
+        Softskill skill = skillService.getOne(id).get();
+        
+        skill.setNombre(dtoSoftskill.getNombre());
+        skill.setNivel(dtoSoftskill.getNivel());
         
         skillService.saveSoftskill(skill);
         
-        return skill;
+        return new ResponseEntity(new Mensaje("Softskill actualizada"), HttpStatus.OK);
     }
     
     @GetMapping ("/skill/encontrar/{id}")
     public Softskill findSoftskill(@PathVariable Long id){
         return skillService.findSoftskill(id);
+    }
+    
+    @GetMapping("/skill/detail/{id}")
+    public ResponseEntity<Softskill> getById(@PathVariable("id") Long id){
+        
+        if(!skillService.existsById(id))
+            return new ResponseEntity(new Mensaje("No existe"), HttpStatus.NOT_FOUND);
+        
+        Softskill softskill = skillService.getOne(id).get();
+        
+        return new ResponseEntity(softskill, HttpStatus.OK);
     }
     
 }
